requirements.txt
openai
discord.py
gTTS
python-dotenv
pycoingecko
